Move flet process handling code to backgound_flet.py
